Refactor project structure
Move components to a separate root source directory
