+++ src/components/diagram/visual_components/Compartment.tsx
