+++ src/components/diagram/visual_components/NodeShape.tsx
