+++ src/components/diagram/visual_components/minimap/index.tsx
